Surpressing signed/unsigned warnings.

+++ LambdaButtonStrip.cpp
@@ -82,7 +82,7 @@ juce::ApplicationCommandTarget* LambdaButtonStrip::getNextCommandTarget()
 
 void LambdaButtonStrip::getAllCommands(Array<CommandID>& commands)
 {
-	for (int i = 0; i < buttons_.size(); i++) {
+	for (size_t i = 0; i < buttons_.size(); i++) {
 		commands.add(commandBaseIndex_ + i);
 	}
 }
